update Category component to properly render CategoryElement
