+++ src/components/Category/index.tsx
@@ -1,14 +1,15 @@
 import { useContext, useEffect, useState } from "react";
-import { Link } from "react-router-dom";
+// import { Link } from "react-router-dom";
 import {
-  currencyFormatter,
+  // currencyFormatter,
   findCategory,
-  findItemBySku,
+  // findItemBySku,
 } from "../../utils/utils";
 import { StyledCategory } from "../styles/Category.styled";
 import Error from "../Error/index";
 import { AppContextInterface, CartContext } from "../../contexts/context";
-import { linkStyle } from "../Categories";
+// import { linkStyle } from "../Categories";
+import CategoryElement from "../CategoryElement";
 
 interface ICategory {
   id: string;
@@ -70,72 +71,34 @@ const Category = ({ id }: ICategory): JSX.Element => {
     }
   };
 
-  const handleAddToCart = (event: any) => {
-    const sku = event.target.value;
-    const item = findItemBySku(sku, itemList);
-    !cart.map((e: any) => e.id).includes(parseInt(sku)) &&
-      setCart([
-        ...cart,
-        {
-          id: item.sku,
-          name: item.name,
-          image: item.image,
-          price: parseInt(item.regularPrice),
-          quantity: 1,
-          category: id,
-          modelNumber: item.modelNumber,
-        },
-      ]);
-  };
+  // const handleAddToCart = (event: any) => {
+  //   const sku = event.target.value;
+  //   const item = findItemBySku(sku, itemList);
+  // /  !cart.map((e: any) => e.id).includes(parseInt(sku)) &&
+  //     setCart([
+  //       ...cart,
+  //       {
+  //         id: item.sku,
+  //         name: item.name,
+  //         image: item.image,
+  //         price: parseInt(item.regularPrice),
+  //         quantity: 1,
+  //         category: id,
+  //         modelNumber: item.modelNumber,
+  //       },
+  //     ]);
+  // };
 
   const allItems = itemList
     .filter((item: Item) => !item.name.includes("/"))
     .map((item: Item) => {
       return (
-        <div
+        <CategoryElement
           key={item.sku}
-          id={item.regularPrice}
-          data-cy="item-card"
-          className="item-card"
-        >
-          <div className="image-third">
-            <Link to={`/${id}/${item.name}`} style={linkStyle}>
-              <img className="item-image" src={item.image} alt={item.name} />
-            </Link>
-          </div>
-          <div className="item-info">
-            <Link to={`/${id}/${item.name}`} style={linkStyle}>
-              <h2 className="item-name" data-cy="item-name">
-                {item.name}
-              </h2>
-            </Link>
-            <div className="additional-item-info">
-              <Link to={`/${id}/${item.name}`} style={linkStyle}>
-                <h3 className="item-label">
-                  Model:<span className="item-span"> {item.modelNumber}</span>{" "}
-                </h3>
-              </Link>
-              <Link to={`/${id}/${item.name}`} style={linkStyle}>
-                <h3 className="item-label">
-                  SKU:<span className="item-span"> {item.sku}</span>{" "}
-                </h3>
-              </Link>
-            </div>
-          </div>
-          <div className="price-cart">
-            <h3 className="item-price" data-cy="price">
-              {currencyFormatter.format(parseInt(item.regularPrice))}
-            </h3>
-            <button
-              data-cy="add-to-cart-button"
-              value={item.sku}
-              className={`cart-button ${item.sku}`}
-              onClick={handleAddToCart}
-            >
-              Add to Cart
-            </button>
-          </div>
-        </div>
+          itemList={itemList}
+          item={item}
+          id={id}
+        />
       );
     });
